Enable `-Werror` and a reasonable set of warnings
Instead of using `-Weverything` and turning warnings off as they become
problematic, enable `-Werror` and turn on a specific set of warnings
judged to have high signal-to-noise ratio. Fix up all instances of those
warnings in the code.

These warnings are all available in GHC 8.0.1, which is the default in
the current Debian (Debian 9).